refactor: organize questionnaire constants by concern
